requirements.txt
click
tabulate
python-dotenv